New feature to optimize delay during mapping.

+++ src/map/if/ifTime.c
@@ -205,17 +205,17 @@ void If_CutPropagateRequired( If_Man_t * p, If_Obj_t * pObj, If_Cut_t * pCut, fl
             if ( p->pPars->fDelayOpt )
             {
                 int Delay = If_CutSopBalancePinDelays( p, pCut, pPerm );
-                assert( Delay == pCut->Delay );
+                assert( Delay == (int)pCut->Delay );
             }
             else if ( p->pPars->fDelayOptLut )
             {
                 int Delay = If_CutLutBalancePinDelays( p, pCut, pPerm );
-                assert( Delay == pCut->Delay );
+                assert( Delay == (int)pCut->Delay );
             }
             else if ( p->pPars->fDsdBalance )
             {
                 int Delay = If_CutDsdBalancePinDelays( p, pCut, pPerm );
-                assert( Delay == pCut->Delay );
+                assert( Delay == (int)pCut->Delay );
             }
             else 
                 pPerm = pCut->pPerm;
